CSVFile
If the client isn't connected to the server and wants to add a time slot. The data is written to a csv file. Note: The structure of the data isn't complete and the output isn't correct yet. If client is connected then we can send a string of data separated by commas.