Actual status update
Module name is now correctly displayed.
I decided to remove the whole class ordeal but I left some stuff in case I want to get back to it.
For the updates, we can now display a timetable in UI and although it doesn't update automatically, that's the only thing left to do. Currently, you can add or remove classes from the timetable and everytime you open the timetible it will show it the way it's supposed to be.
There is a way of recognising when we have added a new entry to a day but still have TODO recognising what entry gets removed :+1: